Add the default reward calculator
