+++ pallets/parachain-staking/src/mock.rs
@@ -22,18 +22,18 @@
 use super::*;
 use crate::{
 	reward_rate::RewardRateInfo,
-	types::{BalanceOf, Candidate, Reward},
+    reward_config_calc::{RewardRateConfigTrait, DefaultRewardCalculator},
 	{self as stake},
 };
 use frame_support::{
 	construct_runtime, parameter_types,
 	traits::{Currency, GenesisBuild, OnFinalize, OnInitialize},
 	weights::Weight,
-	BoundedVec, PalletId,
+	PalletId,
 };
 use pallet_authorship::EventHandler;
 use sp_consensus_aura::sr25519::AuthorityId;
-use sp_core::{Get, H256};
+use sp_core::H256;
 use sp_runtime::{
 	impl_opaque_keys,
 	testing::{Header, UintAuthorityId},
@@ -148,74 +148,80 @@ parameter_types! {
 	pub const MaxUnstakeRequests: u32 = 6;
 }
 
-pub trait RewardRateConfigTrait {
-	fn reward_rate_config() -> RewardRateInfo;
-}
+/*
+ * pub trait RewardRateConfigTrait {
+ *     fn reward_rate_config() -> RewardRateInfo;
+ * }
+ */
 
-pub struct MockRewardCalculator<T: Config + RewardRateConfigTrait> {
-	_phantom: PhantomData<T>,
-}
+/*
+ * pub struct MockRewardCalculator<T: Config + RewardRateConfigTrait> {
+ *     _phantom: PhantomData<T>,
+ * }
+ */
 
-// [TODO] It's the same as the fix presentage's implmentation
-impl<T: Config + RewardRateConfigTrait> CollatorDelegatorBlockRewardCalculator<T>
-	for MockRewardCalculator<T>
-{
-	fn collator_reward_per_block(
-		stake: &Candidate<T::AccountId, BalanceOf<T>, T::MaxDelegatorsPerCollator>,
-		issue_number: BalanceOf<T>,
-	) -> (Weight, Weight, Reward<T::AccountId, BalanceOf<T>>) {
-		let min_delegator_stake = T::MinDelegatorStake::get();
-		let delegator_sum = (&stake.delegators)
-			.into_iter()
-			.filter(|x| x.amount >= min_delegator_stake)
-			.fold(T::CurrencyBalance::from(0u128), |acc, x| acc + x.amount);
-
-		if delegator_sum == T::CurrencyBalance::from(0u128) {
-			(
-				Weight::from_ref_time(1_u64),
-				Weight::from_ref_time(1_u64),
-				Reward { owner: stake.id.clone(), amount: issue_number },
-			)
-		} else {
-			let collator_reward =
-				T::reward_rate_config().compute_collator_reward::<T>(issue_number);
-			(
-				Weight::from_ref_time(1_u64),
-				Weight::from_ref_time(1_u64),
-				Reward { owner: stake.id.clone(), amount: collator_reward },
-			)
-		}
-	}
-
-	fn delegator_reward_per_block(
-		stake: &Candidate<T::AccountId, BalanceOf<T>, T::MaxDelegatorsPerCollator>,
-		issue_number: BalanceOf<T>,
-	) -> (Weight, Weight, BoundedVec<Reward<T::AccountId, BalanceOf<T>>, T::MaxDelegatorsPerCollator>)
-	{
-		let min_delegator_stake = T::MinDelegatorStake::get();
-		let delegator_sum = (&stake.delegators)
-			.into_iter()
-			.filter(|x| x.amount >= min_delegator_stake)
-			.fold(T::CurrencyBalance::from(0u128), |acc, x| acc + x.amount);
-
-		let inner = (&stake.delegators)
-			.into_iter()
-			.filter(|x| x.amount >= min_delegator_stake)
-			.map(|x| {
-				let staking_rate = Perquintill::from_rational(x.amount, delegator_sum);
-				let delegator_reward = T::reward_rate_config()
-					.compute_delegator_reward::<T>(issue_number, staking_rate);
-				Reward { owner: x.owner.clone(), amount: delegator_reward }
-			})
-			.collect::<Vec<Reward<T::AccountId, BalanceOf<T>>>>();
-
-		(
-			Weight::from_ref_time(1_u64 + 4_u64),
-			Weight::from_ref_time(inner.len() as u64),
-			inner.try_into().expect("Did not extend vec q.e.d."),
-		)
-	}
-}
+/*
+ * // [TODO] It's the same as the fix presentage's implmentation
+ * impl<T: Config + RewardRateConfigTrait> CollatorDelegatorBlockRewardCalculator<T>
+ *     for MockRewardCalculator<T>
+ * {
+ *     fn collator_reward_per_block(
+ *         stake: &Candidate<T::AccountId, BalanceOf<T>, T::MaxDelegatorsPerCollator>,
+ *         issue_number: BalanceOf<T>,
+ *     ) -> (Weight, Weight, Reward<T::AccountId, BalanceOf<T>>) {
+ *         let min_delegator_stake = T::MinDelegatorStake::get();
+ *         let delegator_sum = (&stake.delegators)
+ *             .into_iter()
+ *             .filter(|x| x.amount >= min_delegator_stake)
+ *             .fold(T::CurrencyBalance::from(0u128), |acc, x| acc + x.amount);
+ *
+ *         if delegator_sum == T::CurrencyBalance::from(0u128) {
+ *             (
+ *                 Weight::from_ref_time(1_u64),
+ *                 Weight::from_ref_time(1_u64),
+ *                 Reward { owner: stake.id.clone(), amount: issue_number },
+ *             )
+ *         } else {
+ *             let collator_reward =
+ *                 T::reward_rate_config().compute_collator_reward::<T>(issue_number);
+ *             (
+ *                 Weight::from_ref_time(1_u64),
+ *                 Weight::from_ref_time(1_u64),
+ *                 Reward { owner: stake.id.clone(), amount: collator_reward },
+ *             )
+ *         }
+ *     }
+ *
+ *     fn delegator_reward_per_block(
+ *         stake: &Candidate<T::AccountId, BalanceOf<T>, T::MaxDelegatorsPerCollator>,
+ *         issue_number: BalanceOf<T>,
+ *     ) -> (Weight, Weight, BoundedVec<Reward<T::AccountId, BalanceOf<T>>, T::MaxDelegatorsPerCollator>)
+ *     {
+ *         let min_delegator_stake = T::MinDelegatorStake::get();
+ *         let delegator_sum = (&stake.delegators)
+ *             .into_iter()
+ *             .filter(|x| x.amount >= min_delegator_stake)
+ *             .fold(T::CurrencyBalance::from(0u128), |acc, x| acc + x.amount);
+ *
+ *         let inner = (&stake.delegators)
+ *             .into_iter()
+ *             .filter(|x| x.amount >= min_delegator_stake)
+ *             .map(|x| {
+ *                 let staking_rate = Perquintill::from_rational(x.amount, delegator_sum);
+ *                 let delegator_reward = T::reward_rate_config()
+ *                     .compute_delegator_reward::<T>(issue_number, staking_rate);
+ *                 Reward { owner: x.owner.clone(), amount: delegator_reward }
+ *             })
+ *             .collect::<Vec<Reward<T::AccountId, BalanceOf<T>>>>();
+ *
+ *         (
+ *             Weight::from_ref_time(1_u64 + 4_u64),
+ *             Weight::from_ref_time(inner.len() as u64),
+ *             inner.try_into().expect("Did not extend vec q.e.d."),
+ *         )
+ *     }
+ * }
+ */
 
 impl Config for Test {
 	type RuntimeEvent = RuntimeEvent;
@@ -238,7 +244,7 @@ impl Config for Test {
 	type MaxUnstakeRequests = MaxUnstakeRequests;
 	type PotId = PotId;
 	type WeightInfo = ();
-	type BlockRewardCalculator = MockRewardCalculator<Self>;
+	type BlockRewardCalculator = DefaultRewardCalculator<Self>;
 }
 
 impl RewardRateConfigTrait for Test {
